requirements.txt
# ===========================================
# AI Agents Book - 통합 Requirements
# ===========================================

# -----------------------------------------
# LLM Provider SDKs
# -----------------------------------------
openai>=1.0.0
anthropic>=0.18.0

# -----------------------------------------
# LangChain Framework (Chapter 3, 4, 5, 6)
# -----------------------------------------
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# -----------------------------------------
# LangGraph (Chapter 3, 5, 6)
# -----------------------------------------
langgraph>=0.2.0

# -----------------------------------------
# CrewAI (Chapter 5, 6)
# -----------------------------------------
crewai>=0.67.0

# -----------------------------------------
# 토큰화 및 임베딩 (Chapter 2, 4)
# -----------------------------------------
tiktoken>=0.5.0
sentence-transformers>=2.2.2

# -----------------------------------------
# 검색 (Chapter 4)
# -----------------------------------------
rank-bm25>=0.2.2
duckduckgo-search>=6.0.0

# -----------------------------------------
# 데이터 검증 및 모델링
# -----------------------------------------
pydantic>=2.8.0

# -----------------------------------------
# 웹 프레임워크 및 HTTP (Chapter 6, 7)
# -----------------------------------------
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0
websockets>=12.0
aiohttp>=3.9.0

# -----------------------------------------
# 시스템 모니터링 (Chapter 8, 9)
# -----------------------------------------
psutil>=5.9.0

# -----------------------------------------
# 알림 및 통합 (Chapter 8)
# -----------------------------------------
slack-sdk>=3.27.0

# -----------------------------------------
# 로컬 LLM - Ollama (Chapter 2, 9)
# -----------------------------------------
ollama>=0.3.0

# -----------------------------------------
# 딥러닝 (선택적 - GPU 가속용)
# -----------------------------------------
# torch>=2.0.0  # GPU 사용 시 주석 해제

# -----------------------------------------
# 다국어 처리 (Chapter 9)
# -----------------------------------------
langdetect>=1.0.9

# -----------------------------------------
# 유틸리티
# -----------------------------------------
python-dotenv>=1.0.0
numpy>=1.24.0
